+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [1.4.0] - 2019-09-17
+### Added
+- Update expenses category
+
+### Fixed
+- Mobile responsive design in expenses list
+
 ## [1.3.0] - 2019-09-16
 ### Added
 - Categories